Updating/deleting personalization data
Addresses issue #42

As explained in the arch draft, personalization data need not be
bundled with the TA binary, and can be expressed by a separate
SUIT manifest.  As such, the TEEP protocol must be able to update
and delete such data.

This PR makes no changes to the actual protocol format, it just
updates text in the doc to explicitly allow it.

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>