Address Russ's comments on draft -05
Addresses issue #132 except that it does not yet address the comment
"In Section 9, please say whether future registrations will allow
integrity-without-confidentiality ciphersuites. Let's settle this now
instead of dumping on the IANA Expert."

Will cover that in a separate PR after discussion in TEEP meeting.
All the others are straightforward.

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>