Add unneeded-tc-list back into Update message
It was originally removed in PR #118.
But issue #238 was then encountered during implementation.
This RP restores the field in the hopes that either:
a) the SUIT manifest will allow unlink directions in an install
manifest (option 2-2 in PR #169, and now tracked in
https://github.com/bremoran/suit-multiple-trust-domains/issues/3), or
b) a TEEP Agent can somehow figure out how to uninstall an installed component
   (option 2-1 in PR #169)

Signed-off-by: Dave Thaler <dthaler@microsoft.com>