Making draft-ietf-suit-firmware-encryption normative
draft-ietf-suit-firmware-encryption is the mechanism to encrypt payloads when using the SUIT manifest. I think draft-ietf-suit-firmware-encryption therefore needs to be normative if we want to encrypt TA binaries and personalization data. For the encryption of EATs and SUIT reports draft-ietf-suit-firmware-encryption is not used but instead the underlying primitives are used directly.